Working (but not friendly) image previews
The biggest problem is that settings for image previews are still
context settings, and that is a problem because the matching them
precisely should not be required.

Now if user switches the data set to a one that has different column
names, preprocessors are going to be forgotten. Because the actual
preprocessing is more important than the previews, the settings here
should not be context settings, or should be at least made optional.